cod_poc/ciandt_it-conversational-agents-2c8392d72d36/pyproject.toml
[project]
name = "conversational-agents"
version = "0.1.0"
description = "Multi-agent conversational framework — domain-agnostic, channel-extensible"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=1.0.0,<2.0",
    "strands-agents-tools>=0.1.0,<1.0",
    "bedrock-agentcore>=1.0.0,<2.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "boto3>=1.35.0,<2.0",
    "httpx>=0.27.0,<1.0",
    "python-dotenv>=1.0.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "chainlit>=1.0.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "moto[all]>=5.0",
    "mypy>=1.10",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "S", "B", "UP", "SIM"]
ignore = [
    "S101",  # assert — acceptable in tests
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "unit: Unit tests (nível 1 — zero AWS)",
    "integration: Integration tests com mocks (nível 1)",
    "critical: Critical issues C1-C7",
    "container: Container tests (nível 2)",
    "staging: E2E tests contra staging (nível 3)",
]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=25 -q"
